+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   </summary>
 
   ### Minor
+  * Masonry: Promotes ExperimentalMasonry to be Masonry. Complete re-write of
+  measuring etc. (#101)
+  * Internal: Gestalt now is React 16.2.0 compatible. (#101)
 
   ### Patch